move position model to module folder
