+++ app/Modules/Bookings/BookingsServiceInterface.php
@@ -2,7 +2,7 @@
 
 namespace App\Modules\Bookings;
 
-use App\Position;
+use App\Modules\Position\Position;
 use App\User;
 use Carbon\Carbon;
