+++ database/factories/BookingFactory.php
@@ -1,6 +1,7 @@
 <?php
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
+
 use App\Modules\Bookings\Booking;
 use Faker\Generator as Faker;
 
@@ -10,7 +11,7 @@ $factory->define(Booking::class, function (Faker $faker) {
 
     return [
         'user_id' => $faker->numberBetween(1000000, 9999999),
-        'position_id' => factory(\App\Position::class),
+        'position_id' => factory(\App\Modules\Position\Position::class),
         'from' => $from,
         'to' => $to,
     ];
