+++ tests/Unit/Endorsement/SpecialEndorsementUnitTest.php
@@ -3,7 +3,7 @@
 namespace Tests\Unit\Endorsement;
 
 use App\Modules\Endorsement\Special\SpecialEndorsement;
-use App\Position;
+use App\Modules\Position\Position;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
